Update and bug fixes!

+++ core/VideoCommunication.java
@@ -14,15 +14,15 @@ public class VideoCommunication extends InteractiveCommunication implements Seri
         _to.setBusy();
     }
 
-    protected double computeCost(TariffPlan plan) {
+    protected long computeCost(TariffPlan plan) {
         if(getFrom().checkFriends(getTo())) {
-            if (plan.getName().equals("NORMAL")) return (30.00 * this.getDuration())/2;
-            else if (plan.getName().equals("GOLD")) return (20.00 * this.getDuration())/2;
-            else return (10.00 * this.getDuration())/2;
+            if (plan.getName().equals("NORMAL")) return (30 * this.getDuration())/2;
+            else if (plan.getName().equals("GOLD")) return (20 * this.getDuration())/2;
+            else return (10 * this.getDuration())/2;
         }
-        if (plan.getName().equals("NORMAL")) return 30.00 * this.getDuration();
-        else if (plan.getName().equals("GOLD")) return 20.00 * this.getDuration();
-        else return 10.00 * this.getDuration();
+        if (plan.getName().equals("NORMAL")) return 30 * this.getDuration();
+        else if (plan.getName().equals("GOLD")) return 20 * this.getDuration();
+        else return 10 * this.getDuration();
     }
 
     private void updateDebt(TariffPlan plan){
